refactor: remove util
