+++ src/import/converters.rs
@@ -1,12 +1,10 @@
 // Copyright 2024 the Velato Authors
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-use std::collections::HashMap;
-
 use super::builders::{setup_layer_base, setup_precomp_layer, setup_shape_layer};
 use super::defaults::{FLOAT_VALUE_ONE_HUNDRED, FLOAT_VALUE_ZERO, MULTIDIM_ONE, POSITION_ZERO};
-use crate::import::util::calc_stops;
 use crate::runtime::model::animated::{self, Position};
+use crate::runtime::model::Easing;
 use crate::runtime::model::{
     self, Content, Draw, EasingHandle, GroupTransform, Layer, SplineToPath, Time, Tween, Value,
 };
@@ -19,6 +17,7 @@ use crate::schema::animated_properties::split_vector::SplitVector;
 use crate::schema::constants::gradient_type::GradientType;
 use crate::schema::helpers::int_boolean::BoolInt;
 use crate::{schema, Composition};
+use std::collections::HashMap;
 use vello::kurbo::{Cap, Join, Point, Size, Vec2};
 use vello::peniko::{BlendMode, Color, Mix};
 
@@ -333,7 +332,7 @@ fn conv_gradient_colors(
     match &value.colors.animated_property.value {
         Static(value) => runtime::model::ColorStops::Fixed({
             let mut stops = runtime::model::fixed::ColorStops::new();
-            let raw = calc_stops(value, count);
+            let raw = conv_stops(value, count);
             for values in raw {
                 stops.push(
                     (
@@ -364,7 +363,7 @@ fn conv_gradient_colors(
                     hold,
                 });
 
-                let stops = calc_stops(&value.value, count)
+                let stops = conv_stops(&value.value, count)
                     .into_iter()
                     .flatten()
                     .collect::<Vec<_>>();
@@ -775,4 +774,62 @@ pub fn conv_size(value: &MultiDimensional) -> Value<Size> {
             x.get(1).copied().unwrap_or(0.0),
         )
     })
+}
+
+pub fn conv_stops(value: &[f64], count: usize) -> Vec<[f64; 5]> {
+    let mut stops: Vec<[f64; 5]> = Vec::new();
+    let mut alpha_stops: Vec<(f64, f64)> = Vec::new();
+    for chunk in value.chunks_exact(4) {
+        stops.push([chunk[0], chunk[1], chunk[2], chunk[3], 1.0]);
+        if stops.len() >= count {
+            // there is alpha data at the end of the list, which is a sequence
+            // of (offset, alpha) pairs
+            for chunk in value.chunks_exact(2).skip(count * 2) {
+                let offset = chunk[0];
+                let alpha = chunk[1];
+                alpha_stops.push((offset, alpha));
+            }
+
+            for stop in stops.iter_mut() {
+                let mut last: Option<(f64, f64)> = None;
+                for &(b, alpha_b) in alpha_stops.iter() {
+                    if let Some((a, alpha_a)) = last.take() {
+                        let x = stop[0];
+                        let t = normalize_to_range(a, b, x);
+
+                        let alpha_interp = alpha_a.tween(&alpha_b, t, &Easing::LERP);
+                        let alpha_interp = if (x >= a && x <= b) && (t <= 0.25) && (x <= 0.1) {
+                            alpha_a
+                        } else {
+                            alpha_interp
+                        }; // todo: this is a hack to get alpha rendering with a
+                           // falloff similar to lottiefiles'
+
+                        let alpha_interp = if (x >= a && x <= b) && (t >= 0.75) && (x >= 0.9) {
+                            alpha_b
+                        } else {
+                            alpha_interp
+                        }; // todo: this is a hack to get alpha rendering with a
+                           // falloff similar to lottiefiles'
+
+                        stop[4] = stop[4].min(alpha_interp);
+                    }
+                    last = Some((b, alpha_b));
+                }
+            }
+            break;
+        }
+    }
+
+    stops
+}
+
+pub fn normalize_to_range(a: f64, b: f64, x: f64) -> f64 {
+    if a == b {
+        // Avoid division by zero if a and b are the same
+        return 0.0;
+    }
+
+    // Calculate the normalized value
+    (x - a) / (b - a)
 }
